fix communication bug, add support for polling all of module's receivers, change the way rays are stored
fix outstanding bug with communication, making it finally stable as
it should be. The problem was using one too many write, that was not
removed after copying the function to read all data.

add support for 'p' type polling, which reads all of module's
receivers at once and providing faster communication.

now rays are stored per each sender, making it easy to retain them
for displaying with 'persistence' and to manage them later on. this
makes it easy to see any objects blocking the path of rays (which are
where rays are densly drawn) or faulty receivers (which are drawn
when there's nothing in the frame)